requirements.txt
paramiko
slack-sdk
python-dotenv
azure-storage-blob
httpx
pandas
